requirements.txt
fastapi
uvicorn
python-dotenv
ollama
pytest
pytest-asyncio
httpx
